refactor: restructurings
