+++ src/main/java/zone/vao/nexoAddon/NexoAddon.java
@@ -21,8 +21,8 @@ import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scheduler.BukkitTask;
 import org.jetbrains.annotations.NotNull;
-import zone.vao.nexoAddon.classes.Components;
-import zone.vao.nexoAddon.classes.Mechanics;
+import zone.vao.nexoAddon.components.Components;
+import zone.vao.nexoAddon.mechanics.Mechanics;
 import zone.vao.nexoAddon.commands.NexoAddonCommand;
 import zone.vao.nexoAddon.events.EntityDeathListener;
 import zone.vao.nexoAddon.events.PlayerCommandPreprocessListener;
@@ -46,8 +46,8 @@ import zone.vao.nexoAddon.events.player.PlayerMovementListener;
 import zone.vao.nexoAddon.handlers.BlockHardnessHandler;
 import zone.vao.nexoAddon.handlers.ParticleEffectManager;
 import zone.vao.nexoAddon.handlers.RecipeManager;
-import zone.vao.nexoAddon.hooks.PacketEventsHook;
-import zone.vao.nexoAddon.metrics.Metrics;
+import zone.vao.nexoAddon.utils.hooks.PacketEventsHook;
+import zone.vao.nexoAddon.utils.metrics.Metrics;
 import zone.vao.nexoAddon.populators.CustomChunkGenerator;
 import zone.vao.nexoAddon.populators.orePopulator.CustomOrePopulator;
 import zone.vao.nexoAddon.populators.orePopulator.Ore;
