--- src/main/java/zone/vao/nexoAddon/classes/component/SkullValue.java
@@ -1,4 +1,0 @@
-package zone.vao.nexoAddon.classes.component;
-
-public record SkullValue(String value) {
-}
