+++ src/main/java/zone/vao/nexoAddon/components/Fertilizer.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.component;
+package zone.vao.nexoAddon.components;
 
 import lombok.Getter;
