+++ src/main/java/zone/vao/nexoAddon/components/JukeboxPlayable.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.component;
+package zone.vao.nexoAddon.components;
 
 public record JukeboxPlayable(String songKey) {
 }
