+++ src/main/java/zone/vao/nexoAddon/components/SkullValue.java
@@ -1,0 +1,4 @@
+package zone.vao.nexoAddon.components;
+
+public record SkullValue(String value) {
+}
