+++ src/main/java/zone/vao/nexoAddon/events/nexo/blocks/breaks/BlockAuraListener.java
@@ -3,7 +3,7 @@ package zone.vao.nexoAddon.events.nexo.blocks.breaks;
 import com.nexomc.nexo.api.events.custom_block.NexoBlockBreakEvent;
 import org.bukkit.Location;
 import zone.vao.nexoAddon.NexoAddon;
-import zone.vao.nexoAddon.classes.Mechanics;
+import zone.vao.nexoAddon.mechanics.Mechanics;
 import zone.vao.nexoAddon.utils.BlockUtil;
 
 public class BlockAuraListener {
