+++ src/main/java/zone/vao/nexoAddon/events/nexo/blocks/interacts/ShiftBlockListener.java
@@ -7,7 +7,7 @@ import com.nexomc.nexo.mechanics.custom_block.CustomBlockMechanic;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import zone.vao.nexoAddon.NexoAddon;
-import zone.vao.nexoAddon.classes.Mechanics;
+import zone.vao.nexoAddon.mechanics.Mechanics;
 
 import static zone.vao.nexoAddon.utils.BlockUtil.startShiftBlock;
