+++ src/main/java/zone/vao/nexoAddon/events/nexo/furnitures/interacts/ShiftBlockListener.java
@@ -7,7 +7,7 @@ import com.nexomc.nexo.mechanics.furniture.FurnitureMechanic;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import zone.vao.nexoAddon.NexoAddon;
-import zone.vao.nexoAddon.classes.Mechanics;
+import zone.vao.nexoAddon.mechanics.Mechanics;
 
 import java.util.ArrayList;
 import java.util.Collections;
