+++ src/main/java/zone/vao/nexoAddon/mechanics/BedrockBreak.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.mechanic;
+package zone.vao.nexoAddon.mechanics;
 
 public record BedrockBreak(int hardness, double probability, int durabilityCost, boolean disableOnFirstLayer) {
 }
