+++ src/main/java/zone/vao/nexoAddon/mechanics/BottledExp.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.mechanic;
+package zone.vao.nexoAddon.mechanics;
 
 public record BottledExp(Double ratio, int cost) {
