+++ src/main/java/zone/vao/nexoAddon/mechanics/DropExperience.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.mechanic;
+package zone.vao.nexoAddon.mechanics;
 
 
 public record DropExperience(double experience) {
