+++ src/main/java/zone/vao/nexoAddon/mechanics/MiningTools.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.mechanic;
+package zone.vao.nexoAddon.mechanics;
 
 import org.bukkit.Material;
