+++ src/main/java/zone/vao/nexoAddon/utils/metrics/Metrics.java
@@ -12,7 +12,7 @@
  *
  * Violations will result in a ban of your plugin and account from bStats.
  */
-package zone.vao.nexoAddon.metrics;
+package zone.vao.nexoAddon.utils.metrics;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.YamlConfiguration;
