+++ src/main/resources/config.yml
@@ -1,5 +1,5 @@
 boss_bar: false                             # Boss bar when looking at furniture
-furniture_evolution_status: true            # Display hologram on interacting with Furniture with evolution mechanic
+furniture_evolution_status: true            # Display hologram on interacting with Furniture with evolution mechanics
 prevent_furniture_break_with_tools: true
 prevent_furniture_break_with_sword: true
 double_hit_destroy_mechanic: true           # Destroy furniture with double hit
